renamed to cToken
